djangoauthapi1/requirements.txt
absl-py==2.0.0
asgiref==3.5.0
astunparse==1.6.3
cachetools==5.3.1
catboost==1.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-mongodb-engine==0.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.5.0
earthengine-api==0.1.371
ee==0.2
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.5.4
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpcio==1.60.0
h5py==3.10.0
httplib2==0.22.0
idna==3.4
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mongo==0.2.0
mongoengine==0.27.0
mysqlclient==2.2.0
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.2.4
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.5
Werkzeug==3.0.1
windows-curses==2.3.1
wrapt==1.14.1
